+++ packages/vscode-extension/src/constants.ts
@@ -7,7 +7,8 @@ export const CONSTANTS = {
     CONNECT_COMMAND: 'neo4j.connect',
     DISCONNECT_COMMAND: 'neo4j.disconnect',
     SWITCH_DATABASE_COMMAND: 'neo4j.switchDatabase',
-    RUN_CYPHER_FILE: 'neo4j.runCypherFile',
+    RUN_CYPHER: 'neo4j.runCypher',
+    CYPHER_FILE_FROM_SELECTION: 'neo4j.cypherFileFromSelection',
   },
   MESSAGES: {
     CONNECTED_MESSAGE: 'Connected to Neo4j.',