+++ packages/vscode-extension/src/cypherRunner.ts
@@ -10,7 +10,6 @@ export default class CypherRunner {
 
   async run(connection: Connection, uri: Uri, input: string) {
     const statements = parseStatementsStrs(input);
-
     const filePath = uri.toString();
 
     if (this.results.has(filePath)) {