+++ packages/language-server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @neo4j-cypher/language-server
 
+## 2.0.0-next.12
+
+### Patch Changes
+
+- d85c1e0: Fixes bug with auto-completions
+- Updated dependencies [bccf518]
+  - @neo4j-cypher/language-support@2.0.0-next.11
+
+## 2.0.0-next.11
+
+### Patch Changes
+
+- Updated dependencies [8760c02]
+  - @neo4j-cypher/language-support@2.0.0-next.10
+
 ## 2.0.0-next.10
 
 ### Patch Changes